% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomod3_class5_Wrap.R
\docType{class}
\name{BIOMOD.wrap.out}
\alias{BIOMOD.wrap.out}
\alias{BIOMOD.wrap.out-class}
\alias{show,BIOMOD.wrap.out-method}
\title{\code{BIOMOD_Wrap()} output object class}
\usage{
\S4method{show}{BIOMOD.wrap.out}(object)
}
\description{
Class returned by \code{\link{BIOMOD_Wrap}}, and used by
\code{\link{BIOMOD_ProjectionWrap}}
}
\section{Slots}{

\describe{
\item{\code{modeling.id}}{a \code{character} corresponding to the name (ID) of the
simulation set}

\item{\code{dir.name}}{a \code{character} corresponding to the modeling folder}

\item{\code{sp.name}}{a \code{character} corresponding to the species name}

\item{\code{expl.var.names}}{a \code{vector} containing names of explanatory
variables}

\item{\code{data.type}}{a \code{character} corresponding to the data type}

\item{\code{models.out}}{a \code{\link{BIOMOD.stored.models.out-class}} object
containing informations from \code{\link{BIOMOD_Modeling}} object}
}}

\author{
Helene Blancheteau
}
