% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomod3-package.R
\docType{package}
\name{biomod3-package}
\alias{biomod3}
\alias{biomod3-package}
\title{biomod3: Biomod2 toolkit}
\description{
Biomod2 Workflow and Multispecies.
}
\author{
\strong{Maintainer}: Maya Gueguen \email{maya.gueguen@univ-grenoble-alpes.fr}

Authors:
\itemize{
  \item Helene Blancheteau \email{helene.blancheteau@univ-grenoble-alpes.fr}
}

}
\keyword{internal}
