% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIOMOD_ProjectionWrap.R
\name{BIOMOD_ProjectionWrap}
\alias{BIOMOD_ProjectionWrap}
\title{Project a range of calibrated species distribution models onto new environment}
\usage{
BIOMOD_ProjectionWrap(
  bm.wrap,
  proj.name,
  new.env,
  new.env.xy = NULL,
  models.chosen = "all",
  metric.binary = NULL,
  metric.filter = NULL,
  compress = TRUE,
  digits = 0,
  build.clamping.mask = TRUE,
  nb.cpu = 1,
  seed.val = NULL,
  ...
)
}
\arguments{
\item{bm.wrap}{a \code{\link{BIOMOD.wrap.out}} object returned by the
\code{\link{BIOMOD_Wrap}} function}

\item{proj.name}{a \code{character} corresponding to the name (ID) of the projection set
(\emph{a new folder will be created within the simulation folder with this name})}

\item{new.env}{A \code{matrix}, \code{data.frame} or
\code{\link[terra:rast]{SpatRaster}} object containing the new
explanatory variables (in columns or layers, with names matching the
variables names given to the \code{\link{BIOMOD_FormatingData}} function to build
\code{bm.wrap}) that will be used to project the species distribution model(s)
\cr \emph{Note that old format from \pkg{raster} are still supported such as
\code{RasterStack} objects. }}

\item{new.env.xy}{(\emph{optional, default} \code{NULL}) \cr
If \code{new.env} is a \code{matrix} or a \code{data.frame}, a 2-columns \code{matrix} or
\code{data.frame} containing the corresponding \code{X} and \code{Y} coordinates that will be
used to project the species distribution model(s)}

\item{models.chosen}{a \code{vector} containing model names to be kept, must be either
\code{all}, \code{single}, \code{ensemble} or a sub-selection of model names that can be obtained with the
\code{\link{get_built_models}} function}

\item{metric.binary}{(\emph{optional, default} \code{NULL}) \cr
A \code{vector} containing evaluation metric names to be used to transform prediction values
into binary values based on models evaluation scores obtained with the
\code{\link{BIOMOD_Modeling}} function. Must be among \code{all} (same evaluation metrics than
those of \code{bm.wrap}) or \code{POD}, \code{FAR}, \code{POFD}, \code{SR}, \code{ACCURACY},
\code{BIAS}, \code{ROC}, \code{TSS}, \code{KAPPA}, \code{OR}, \code{ORSS}, \code{CSI},
\code{ETS}, \code{BOYCE}, \code{MPA}}

\item{metric.filter}{(\emph{optional, default} \code{NULL}) \cr
A \code{vector} containing evaluation metric names to be used to transform prediction values
into filtered values based on models evaluation scores obtained with the
\code{\link{BIOMOD_Modeling}} function. Must be among \code{all} (same evaluation metrics than
those of \code{bm.wrap}) or \code{POD}, \code{FAR}, \code{POFD}, \code{SR}, \code{ACCURACY},
\code{BIAS}, \code{ROC}, \code{TSS}, \code{KAPPA}, \code{OR}, \code{ORSS}, \code{CSI},
\code{ETS}, \code{BOYCE}, \code{MPA}}

\item{compress}{(\emph{optional, default} \code{TRUE}) \cr
A \code{logical} or a \code{character} value defining whether and how objects should be
compressed when saved on hard drive. Must be either \code{TRUE}, \code{FALSE}, \code{xz} or
\code{gzip} (see Details)}

\item{digits}{(\emph{optional, default} \code{0}) \cr
A \code{integer} value defining the number of digits of the predictions.}

\item{build.clamping.mask}{(\emph{optional, default} \code{TRUE}) \cr
A \code{logical} value defining whether a clamping mask should be built and saved on hard
drive or not (see Details)}

\item{nb.cpu}{(\emph{optional, default} \code{1}) \cr
An \code{integer} value corresponding to the number of computing resources to be used to
parallelize the single models computation}

\item{seed.val}{(\emph{optional, default} \code{NULL}) \cr
An \code{integer} value corresponding to the new seed value to be set}

\item{\ldots}{(\emph{optional, see Details)})}
}
\value{
A \code{\link{BIOMOD.projection.out}} object containing models projections, or links to saved
outputs. \cr Models projections are stored out of \R (for memory storage reasons) in
\code{proj.name} folder created in the current working directory :
\enumerate{
\item the output is a \code{data.frame} if \code{new.env} is a \code{matrix} or a
\code{data.frame}
\item it is a \code{\link[terra:rast]{SpatRaster}} if \code{new.env} is a
\code{\link[terra:rast]{SpatRaster}} (or several \code{\link[terra:rast]{SpatRaster}}
objects, if \code{new.env} is too large)
\item raw projections, as well as binary and filtered projections (if asked), are saved in
the \code{proj.name} folder
}
}
\description{
This function allows to project a range of models built with the
\code{\link{BIOMOD_Wrap}} function onto new environmental data (\emph{which can
represent new areas, resolution or time scales for example}).
}
\examples{
library(terra)
library(biomod2)

# Load species occurrences (6 species available)
data(DataSpecies)
myRespName <- 'GuloGulo'
myResp <- as.numeric(DataSpecies[, myRespName])
myRespXY <- DataSpecies[, c('X_WGS84', 'Y_WGS84')]

# Load environmental variables extracted from BIOCLIM (bio_3, bio_4, bio_7, bio_11 & bio_12)
data(bioclim_current)
myExpl <- terra::rast(bioclim_current)

\dontshow{
  myExtent <- terra::ext(0,30,45,70)
  myExpl <- terra::crop(myExpl, myExtent)
}
# ---------------------------------------------------------------#
# Data formating, modeling and ensemble modeling in one workflow
myWrap <- BIOMOD_Wrap(modeling.id = "Example",
                      data.type = "binary",
                      resp.name = myRespName,
                      resp.var = myResp,
                      resp.xy = myRespXY,
                      expl.var = myExpl,
                      models = c('RF', 'GLM'),
                      metric.eval = c("TSS", "ROC"),
                      params.CV = list(CV.strategy = 'random',
                                       CV.nb.rep = 3,
                                       CV.perc = 0.7),
                      em.algo = c('EMmean', 'EMca'),
                      params.EM = list(models.chosen = 'all',
                                       em.by = 'all',
                                       metric.select = 'TSS',
                                       metric.select.thresh = 0.6),
                      var.import = 2,
                      seed.val = 42)

myProjections <- BIOMOD_ProjectionWrap(bm.wrap = myWrap,
                                       proj.name = "Current",
                                       new.env = myExpl,
                                       models.chosen = 'all')
# plot(myProjections)
\dontshow{
  unlink('GuloGulo', recursive = TRUE)
}

}
\author{
Helene Blancheteau
}
