% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIOMOD_Wrap.R
\name{BIOMOD_Wrap}
\alias{BIOMOD_Wrap}
\title{Run a range of species distribution models}
\usage{
BIOMOD_Wrap(
  dir.name = ".",
  modeling.id = as.character(format(Sys.time(), "\%s")),
  data.type = "binary",
  resp.name,
  resp.var,
  resp.xy = NULL,
  expl.var,
  eval.resp.var = NULL,
  eval.resp.xy = NULL,
  eval.expl.var = NULL,
  filter.raster = FALSE,
  params.PA,
  models,
  models.pa = NULL,
  metric.eval = c("KAPPA", "TSS", "ROC"),
  weights = NULL,
  prevalence = NULL,
  scale.models = FALSE,
  var.import = 0,
  params.CV,
  params.OPT,
  em.algo,
  params.EM,
  seed.val = NULL,
  nb.cpu = 1
)
}
\arguments{
\item{dir.name}{(\emph{optional, default} \code{.}) \cr
A \code{character} corresponding to the modeling folder}

\item{modeling.id}{a \code{character} corresponding to the name (ID) of the simulation set
(\emph{a random number by default})}

\item{data.type}{a \code{character}, corresponding to the response data type to be used, must be either
\code{binary}, \code{count}, \code{ordinal}, \code{relative}, or \code{abundance}. If data.type is not provided,
\code{biomod2} will try to guess.}

\item{resp.name}{a \code{character vector} corresponding to the species name}

\item{resp.var}{a \code{vector} or a \code{\link[terra:vect]{SpatVector}} containing your response variable (See Details).}

\item{resp.xy}{(\emph{optional, default} \code{NULL}) \cr
If \code{resp.var} is a \code{vector}, a 2-columns \code{matrix} or \code{data.frame}
containing the corresponding \code{X} and \code{Y} coordinates that will be
used to build the species distribution model(s)}

\item{expl.var}{a \code{matrix}, \code{data.frame}, \code{\link[terra:vect]{SpatVector}}
or \code{\link[terra:rast]{SpatRaster}} object containing the explanatory variables
(in columns or layers) that will be used to build the species distribution model(s)
\cr \emph{Note that old format from \pkg{raster} and \pkg{sp} are still supported such as
\code{RasterStack} and \code{SpatialPointsDataFrame} objects. }}

\item{eval.resp.var}{(\emph{optional, default} \code{NULL}) \cr
A \code{vector}, a \code{\link[terra:vect]{SpatVector}}
without associated data (\emph{if presence-only}),
or a \code{\link[terra:vect]{SpatVector}} object containing binary data
(\code{0} : absence, \code{1} : presence, \code{NA} : indeterminate)
for a single species that will be used to
evaluate the species distribution model(s) with independent data
\cr \emph{Note that old format from \pkg{sp} are still supported such as
\code{SpatialPoints}  (if presence-only) or \code{SpatialPointsDataFrame}
object containing binary data.}}

\item{eval.resp.xy}{(\emph{optional, default} \code{NULL}) \cr
If \code{resp.var} is a \code{vector}, a 2-columns \code{matrix} or \code{data.frame}
containing the corresponding \code{X} and \code{Y} coordinates that will be used to evaluate
the species distribution model(s) with independent data}

\item{eval.expl.var}{(\emph{optional, default} \code{NULL}) \cr
A \code{matrix}, \code{data.frame}, \code{\link[terra:vect]{SpatVector}}
or \code{\link[terra:rast]{SpatRaster}} object containing the explanatory variables (in
columns or layers) that will be used to evaluate the species distribution model(s) with
independent data.
\cr \emph{Note that old format from \pkg{raster} and \pkg{sp} are still supported such as
\code{RasterStack} and \code{SpatialPointsDataFrame} objects. }}

\item{filter.raster}{(\emph{optional, default} \code{FALSE}) \cr
If \code{expl.var} is of raster type, a \code{logical} value defining whether \code{resp.var}
is to be filtered when several points occur in the same raster cell}

\item{params.PA}{a \code{list} with the species names associated to the parameters of PA}

\item{models}{a \code{vector} containing model names to be computed, must be among
\code{ANN}, \code{CTA}, \code{FDA}, \code{GAM}, \code{GBM}, \code{GLM}, \code{MARS},
\code{MAXENT}, \code{MAXNET}, \code{RF}, \code{RFd}, \code{SRE}, \code{XGBOOST}}

\item{models.pa}{(\emph{optional, default} \code{NULL}) \cr
A \code{list} containing for each model a \code{vector} defining which pseudo-absence datasets
are to be used, must be among \code{colnames(bm.format@PA.table)}}

\item{metric.eval}{a \code{vector} containing evaluation metric names to be used, must
be among \code{ROC}, \code{TSS}, \code{KAPPA}, \code{ACCURACY}, \code{BIAS}, \code{POD},
\code{FAR}, \code{POFD}, \code{SR}, \code{CSI}, \code{ETS}, \code{OR},
\code{ORSS}, \code{BOYCE}, \code{MPA}, \code{RMSE}, \code{MAE}, \code{MSE}, \code{Rsquared}, \code{Rsquared_aj},
\code{Max_error}, \code{Accuracy}, \code{Recall}, \code{Precision}, \code{F1}}

\item{weights}{(\emph{optional, default} \code{NULL}) \cr
A \code{vector} of \code{numeric} values corresponding to observation weights (one per
observation, see Details)}

\item{prevalence}{(\emph{optional, default} \code{NULL}) \cr
A \code{numeric} between \code{0} and \code{1} corresponding to the species prevalence to
build '\emph{weighted response weights}' (see Details)}

\item{scale.models}{(\emph{optional, default} \code{FALSE}) \cr
A \code{logical} value defining whether all models predictions should be scaled with a
binomial GLM or not}

\item{var.import}{(\emph{optional, default} \code{NULL}) \cr
An \code{integer} corresponding to the number of permutations to be done for each variable to
estimate variable importance}

\item{params.CV}{a \code{list} with the species names associated to the parameters of Cross-Validation. See BIOMOD_Modeling}

\item{params.OPT}{a \code{list} with the species names associated to the options of the algorithms. See BIOMOD_Modeling}

\item{em.algo}{a \code{vector} corresponding to the ensemble models that will be computed,
must be among \code{'EMmean'}, \code{'EMmedian'}, \code{'EMcv'}, \code{'EMci'},
\code{'EMca'}, \code{'EMwmean'}}

\item{params.EM}{a \code{list} with the species names associated to the parameters for the ensemble models. See BIOMOD_EnsembleModeling}

\item{seed.val}{(\emph{optional, default} \code{NULL}) \cr
An \code{integer} value corresponding to the new seed value to be set}

\item{nb.cpu}{(\emph{optional, default} \code{1}) \cr
An \code{integer} value corresponding to the number of computing resources to be used to
parallelize the single models computation}
}
\value{
A \code{\link{MS.models.out}} object acting as a proxi for the created \code{BIOMOD.models.out}.
}
\description{
This function allows to run a workflow for one or several species using biomod2.
The parameters could be change inside the arguments 'params. 's.
}
\examples{
library(terra)
library(biomod2)

# Load species occurrences (6 species available)
data(DataSpecies)
head(DataSpecies)

# Select the name of the studied species
myRespName <- 'GuloGulo'

# Get corresponding presence/absence data
myResp <- as.numeric(DataSpecies[, myRespName])

# Get corresponding XY coordinates
myRespXY <- DataSpecies[, c('X_WGS84', 'Y_WGS84')]

# Load environmental variables extracted from BIOCLIM (bio_3, bio_4, bio_7, bio_11 & bio_12)
data(bioclim_current)
myExpl <- terra::rast(bioclim_current)

\dontshow{
  myExtent <- terra::ext(0,30,45,70)
  myExpl <- terra::crop(myExpl, myExtent)
}

# ---------------------------------------------------------------#
# Data formating, modeling and ensemble modeling in one workflow
myWrap <- BIOMOD_Wrap(modeling.id = "Example",
                      data.type = "binary",
                      resp.name = myRespName,
                      resp.var = myResp,
                      resp.xy = myRespXY,
                      expl.var = myExpl,
                      models = c('RF', 'GLM'),
                      metric.eval = c("TSS", "ROC"),
                      params.CV = list(CV.strategy = 'random',
                                       CV.nb.rep = 3,
                                       CV.perc = 0.7),
                      em.algo = c('EMmean', 'EMca'),
                      params.EM = list(models.chosen = 'all',
                                       em.by = 'all',
                                       metric.select = 'TSS',
                                       metric.select.thresh = 0.6),
                      var.import = 2,
                      seed.val = 42)

myWrap

\dontshow{
  unlink('GuloGulo', recursive = TRUE)
}


}
\author{
Helene Blancheteau
}
