% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomod3_class2_SM.R, R/biomod3_class3_EM.R,
%   R/biomod3_class4_Proj.R
\name{getters.out}
\alias{getters.out}
\alias{get_species_data}
\alias{get_eval_data}
\alias{get_options}
\alias{get_calib_lines}
\alias{get_formal_data}
\alias{get_projected_models}
\alias{free}
\alias{get_predictions}
\alias{get_kept_models}
\alias{get_built_models}
\alias{get_evaluations}
\alias{get_variables_importance}
\alias{get_options,MS.models.out-method}
\alias{get_calib_lines,MS.models.out-method}
\alias{get_formal_data,MS.models.out-method}
\alias{get_predictions,MS.models.out-method}
\alias{get_built_models,MS.models.out-method}
\alias{get_evaluations,MS.models.out-method}
\alias{get_variables_importance,MS.models.out-method}
\alias{get_formal_data,MS.ensemble.models.out-method}
\alias{get_built_models,MS.ensemble.models.out-method}
\alias{get_kept_models,MS.ensemble.models.out-method}
\alias{get_predictions,MS.ensemble.models.out-method}
\alias{get_evaluations,MS.ensemble.models.out-method}
\alias{get_variables_importance,MS.ensemble.models.out-method}
\alias{get_projected_models,MS.projection.out-method}
\alias{get_predictions,MS.projection.out-method}
\title{Functions to extract informations from \code{\link{MS.models.out}},
\code{\link{MS.projection.out}} or \code{\link{MS.ensemble.models.out}} objects}
\usage{
get_species_data(obj, ...)

\S4method{get_options}{MS.models.out}(obj, sp)

\S4method{get_calib_lines}{MS.models.out}(obj, as.data.frame = FALSE, PA = NULL, run = NULL)

\S4method{get_formal_data}{MS.models.out}(obj, sp, subinfo = NULL)

\S4method{get_predictions}{MS.models.out}(
  obj,
  sp,
  evaluation = FALSE,
  full.name = NULL,
  PA = NULL,
  run = NULL,
  algo = NULL,
  model.as.col = FALSE
)

\S4method{get_built_models}{MS.models.out}(obj, full.name = NULL, PA = NULL, run = NULL, algo = NULL)

\S4method{get_evaluations}{MS.models.out}(
  obj,
  sp,
  full.name = NULL,
  PA = NULL,
  run = NULL,
  algo = NULL,
  metric.eval = NULL
)

\S4method{get_variables_importance}{MS.models.out}(
  obj,
  sp,
  full.name = NULL,
  PA = NULL,
  run = NULL,
  algo = NULL,
  expl.var = NULL
)

\S4method{get_formal_data}{MS.ensemble.models.out}(obj, sp, subinfo = NULL)

\S4method{get_built_models}{MS.ensemble.models.out}(
  obj,
  full.name = NULL,
  merged.by.algo = NULL,
  merged.by.run = NULL,
  merged.by.PA = NULL,
  filtered.by = NULL,
  algo = NULL
)

\S4method{get_kept_models}{MS.ensemble.models.out}(obj)

\S4method{get_predictions}{MS.ensemble.models.out}(
  obj,
  sp,
  evaluation = FALSE,
  full.name = NULL,
  merged.by.algo = NULL,
  merged.by.run = NULL,
  merged.by.PA = NULL,
  filtered.by = NULL,
  algo = NULL,
  model.as.col = FALSE
)

\S4method{get_evaluations}{MS.ensemble.models.out}(
  obj,
  sp,
  full.name = NULL,
  merged.by.algo = NULL,
  merged.by.run = NULL,
  merged.by.PA = NULL,
  filtered.by = NULL,
  algo = NULL,
  metric.eval = NULL
)

\S4method{get_variables_importance}{MS.ensemble.models.out}(
  obj,
  sp,
  full.name = NULL,
  merged.by.algo = NULL,
  merged.by.run = NULL,
  merged.by.PA = NULL,
  filtered.by = NULL,
  algo = NULL,
  expl.var = NULL
)

\S4method{get_projected_models}{MS.projection.out}(
  obj,
  full.name = NULL,
  PA = NULL,
  run = NULL,
  algo = NULL,
  merged.by.algo = NULL,
  merged.by.run = NULL,
  merged.by.PA = NULL,
  filtered.by = NULL
)

\S4method{get_predictions}{MS.projection.out}(
  obj,
  sp,
  metric.binary = NULL,
  metric.filter = NULL,
  full.name = NULL,
  PA = NULL,
  run = NULL,
  algo = NULL,
  merged.by.algo = NULL,
  merged.by.run = NULL,
  merged.by.PA = NULL,
  filtered.by = NULL,
  model.as.col = FALSE,
  ...
)
}
\arguments{
\item{obj}{a \code{\link{MS.formated.data}}, \code{\link{MS.models.out}}, \code{\link{MS.projection.out}} or
\code{\link{MS.ensemble.models.out}} object}

\item{...}{additional parameters}

\item{sp}{a \code{character} name of a specific species}

\item{as.data.frame}{a \code{logical} defining whether output should be returned as
\code{data.frame} or \code{array} object}

\item{PA}{(\emph{optional, default} \code{NULL}) \cr
A \code{vector} containing pseudo-absence set to be loaded, must be among \code{PA1},
\code{PA2}, \code{...}, \code{allData}}

\item{run}{(\emph{optional, default} \code{NULL}) \cr
A \code{vector} containing repetition set to be loaded, must be among \code{RUN1},
\code{RUN2}, \code{...}, \code{allRun}}

\item{subinfo}{a \code{character} corresponding to the information to be extracted, must be
among \code{NULL}, \code{expl.var.names}, \code{resp.var}, \code{expl.var}, \code{MinMax},
\code{eval.resp.var}, \code{eval.expl.var} (see Details)}

\item{evaluation}{a \code{logical} defining whether evaluation data should be used or not}

\item{full.name}{(\emph{optional, default} \code{NULL}) \cr
A \code{vector} containing model names to be kept, must be either \code{all} or a
sub-selection of model names that can be obtained with the \code{\link{get_built_models}}
function}

\item{algo}{(\emph{optional, default} \code{NULL}) \cr
A \code{character} containing algorithm to be loaded, must be either
\code{ANN}, \code{CTA}, \code{FDA}, \code{GAM}, \code{GBM}, \code{GLM}, \code{MARS},
\code{MAXENT}, \code{MAXNET}, \code{RF}, \code{SRE}, \code{XGBOOST}}

\item{model.as.col}{(\emph{optional, default} \code{FALSE}) \cr
A \code{boolean} given to \code{\link{get_predictions}}. If \code{TRUE}
prediction are returned as a wide \code{data.frame} with each column containing
predictions for a single model and corresponding to the old output given by
\pkg{biomod2} in version < 4.2-2. If \code{FALSE} predictions are returned
as a long \code{data.frame} with many additional informations readily
available.}

\item{metric.eval}{(\emph{optional, default} \code{NULL}) \cr
A \code{vector} containing evaluation metric to be kept, must be among \code{POD},
\code{FAR}, \code{POFD}, \code{SR}, \code{ACCURACY}, \code{BIAS}, \code{ROC}, \code{TSS},
\code{KAPPA}, \code{OR}, \code{ORSS}, \code{CSI}, \code{ETS}, \code{BOYCE}, \code{MPA}}

\item{expl.var}{(\emph{optional, default} \code{NULL}) \cr
A \code{vector} containing explanatory variables to be kept, that can be obtained with the
\code{\link{get_formal_data}(obj, subinfo = 'expl.var.names')} function}

\item{merged.by.algo}{(\emph{optional, default} \code{NULL}) \cr
A \code{character} containing merged algorithm to be loaded, must be among
\code{ANN}, \code{CTA}, \code{FDA}, \code{GAM}, \code{GBM}, \code{GLM}, \code{MARS},
\code{MAXENT}, \code{MAXNET}, \code{RF}, \code{SRE}, \code{XGBOOST}, \code{mergedAlgo}}

\item{merged.by.run}{(\emph{optional, default} \code{NULL}) \cr
A \code{vector} containing merged repetition set to be loaded, must be among \code{RUN1},
\code{RUN2}, \code{...}, \code{mergedRun}}

\item{merged.by.PA}{(\emph{optional, default} \code{NULL}) \cr
A \code{vector} containing merged pseudo-absence set to be loaded, must be among \code{PA1},
\code{PA2}, \code{...}, \code{mergedData}}

\item{filtered.by}{(\emph{optional, default} \code{NULL}) \cr
A \code{vector} containing evaluation metric selected to filter single models to build the
ensemble models, must be among \code{POD}, \code{FAR}, \code{POFD}, \code{SR},
\code{ACCURACY}, \code{BIAS}, \code{ROC}, \code{TSS}, \code{KAPPA}, \code{OR}, \code{ORSS},
\code{CSI}, \code{ETS}, \code{BOYCE}, \code{MPA}}

\item{metric.binary}{(\emph{optional, default} \code{NULL}) \cr
A \code{vector} containing evaluation metric selected to transform predictions into binary
values, must be among \code{POD}, \code{FAR}, \code{POFD}, \code{SR}, \code{ACCURACY},
\code{BIAS}, \code{ROC}, \code{TSS}, \code{KAPPA}, \code{OR}, \code{ORSS}, \code{CSI},
\code{ETS}, \code{BOYCE}, \code{MPA}}

\item{metric.filter}{(\emph{optional, default} \code{NULL}) \cr
A \code{vector} containing evaluation metric to filter predictions, must be among \code{POD},
\code{FAR}, \code{POFD}, \code{SR}, \code{ACCURACY}, \code{BIAS}, \code{ROC}, \code{TSS},
\code{KAPPA}, \code{OR}, \code{ORSS}, \code{CSI}, \code{ETS}, \code{BOYCE}, \code{MPA}}
}
\description{
Functions to extract informations from \code{\link{MS.models.out}},
\code{\link{MS.projection.out}} or \code{\link{MS.ensemble.models.out}} objects
}
