% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.Params.R
\name{help.Params}
\alias{help.Params}
\title{Shiny application to help set the parameter for the parameters in
BIOMOD_Wrap or MS functions.}
\usage{
help.Params()
}
\description{
...
}
\author{
Helene Blancheteau
}
