% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomod3_class3_EM.R
\docType{class}
\name{MS.ensemble.models.out}
\alias{MS.ensemble.models.out}
\alias{MS.ensemble.models.out-class}
\alias{show,MS.ensemble.models.out-method}
\title{\code{BIOMOD_EnsembleModeling()} output object class}
\usage{
\S4method{show}{MS.ensemble.models.out}(object)
}
\description{
Class returned by \code{\link{BIOMOD_EnsembleModeling}}, and used by
\code{\link{BIOMOD_LoadModels}}, \code{\link{BIOMOD_PresenceOnly}} and
\code{\link{BIOMOD_EnsembleForecasting}}
}
\section{Slots}{

\describe{
\item{\code{ms.project}}{a \code{character} corresponding to the name of the multispecies project}

\item{\code{modeling.id}}{a \code{character} corresponding to the name (ID) of the
simulation set}

\item{\code{dir.name}}{a \code{character} corresponding to the modeling folder}

\item{\code{sp.name}}{a \code{character} corresponding to the species name}

\item{\code{expl.var.names}}{a \code{vector} containing names of explanatory
variables}

\item{\code{data.type}}{a \code{character} corresponding to the data type}

\item{\code{models.out}}{a \code{\link{BIOMOD.stored.models.out-class}} object
containing informations from \code{\link{BIOMOD_Modeling}} object}

\item{\code{em.by}}{a \code{character} corresponding to the way kept models have
been combined to build the ensemble models, must be among
\code{PA+run}, \code{PA+algo}, \code{PA},
\code{algo}, \code{all}}

\item{\code{em.computed}}{a \code{vector} containing names of ensemble models}

\item{\code{em.failed}}{a \code{vector} containing names of failed ensemble models}

\item{\code{em.models_kept}}{a \code{list} containing single models for each ensemble model}

\item{\code{link}}{a \code{character} containing the file name of the saved object}
}}

\seealso{
\code{\link{BIOMOD_EnsembleModeling}}, \code{\link{BIOMOD_LoadModels}},
\code{\link{BIOMOD_PresenceOnly}}, \code{\link{bm_VariablesImportance}},
\code{\link{bm_PlotEvalMean}}, \code{\link{bm_PlotEvalBoxplot}},
\code{\link{bm_PlotVarImpBoxplot}}, \code{\link{bm_PlotResponseCurves}}

Other Toolbox objects: 
\code{\link{MS.projection.out}}
}
\author{
Damien Georges
}
\concept{Toolbox objects}
