% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomod3_class1_Fd.R
\docType{class}
\name{MS.formated.data}
\alias{MS.formated.data}
\alias{MS.formated.data-class}
\alias{show,MS.formated.data-method}
\title{\code{MS_FormatingData()} output object class}
\usage{
\S4method{show}{MS.formated.data}(object)
}
\arguments{
\item{object}{a \code{\link{MS.formated.data}} object}
}
\description{
Class returned by \code{\link{MS_FormatingData}}
\code{\link{MS_Modeling}}
}
\section{Slots}{

\describe{
\item{\code{data.type}}{a \code{character} corresponding to the data type}

\item{\code{dir.name}}{a \code{character} corresponding to the modeling folder}

\item{\code{ms.project}}{a \code{character} correspondinf to the name of the multispecies project}

\item{\code{sp.name}}{a \code{character vector} corresponding to the species name}

\item{\code{coord}}{a 2-columns \code{data.frame} containing the corresponding \code{X} and \code{Y}
coordinates}

\item{\code{data.species}}{a \code{data.frame} containing the species observations (\code{0}, \code{1} or
\code{NA})}

\item{\code{data.env.var}}{a \code{data.frame} containing explanatory variables}

\item{\code{data.mask}}{a \code{\link[terra:rast]{SpatRaster}} object containing the mask of the
studied area}

\item{\code{PA.strategy}}{a \code{character vector} corresponding to the pseudo-absence selection strategy}

\item{\code{PA.table}}{a \code{list} with the \code{data.frames} containing the corresponding table of selected
pseudo-absences (indicated by \code{TRUE} or \code{FALSE}) from the \code{pa.tab} list
element returned by the \code{\link{bm_PseudoAbsences}} function}

\item{\code{has.data.eval}}{a \code{logical} value defining whether evaluation data is given}

\item{\code{has.filter.raster}}{a \code{logical} value defining whether filtering have been done or not}

\item{\code{eval.coord}}{(\emph{optional, default} \code{NULL}) \cr
A 2-columns \code{data.frame} containing the corresponding \code{X} and \code{Y}
coordinates for evaluation data}

\item{\code{eval.data.species}}{(\emph{optional, default} \code{NULL}) \cr
A \code{vector} containing the species observations (\code{0}, \code{1} or \code{NA}) for
evaluation data}

\item{\code{eval.data.env.var}}{(\emph{optional, default} \code{NULL}) \cr
A \code{data.frame} containing explanatory variables for evaluation data}

\item{\code{biomod3.version}}{a \code{character} corresponding to the biomod3 version}
}}

\author{
Damien Georges
}
