% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomod3_class2_SM.R
\docType{class}
\name{MS.models.out}
\alias{MS.models.out}
\alias{MS.models.out-class}
\alias{getters.out}
\alias{get_species_data}
\alias{get_eval_data}
\alias{get_options}
\alias{get_calib_lines}
\alias{get_formal_data}
\alias{get_projected_models}
\alias{free}
\alias{get_predictions}
\alias{get_kept_models}
\alias{get_built_models}
\alias{get_evaluations}
\alias{get_variables_importance}
\title{\code{MS_Modeling()} output object class}
\usage{
\S4method{show}{MS.models.out}(object)
}
\arguments{
\item{object}{a \code{\link{MS.models.out}} object}

\item{obj}{a \code{\link{MS.formated.data}}, \code{\link{MS.models.out}}, \code{\link{MS.projection.out}} or
\code{\link{MS.ensemble.models.out}} object}

\item{sp}{a \code{character} name of a specific species}

\item{as.data.frame}{a \code{logical} defining whether output should be returned as
\code{data.frame} or \code{array} object}

\item{subinfo}{a \code{character} corresponding to the information to be extracted, must be
among \code{NULL}, \code{expl.var.names}, \code{resp.var}, \code{expl.var}, \code{MinMax},
\code{eval.resp.var}, \code{eval.expl.var} (see Details)}

\item{evaluation}{a \code{logical} defining whether evaluation data should be used or not}

\item{full.name}{(\emph{optional, default} \code{NULL}) \cr
A \code{vector} containing model names to be kept, must be either \code{all} or a
sub-selection of model names that can be obtained with the \code{\link{get_built_models}}
function}

\item{PA}{(\emph{optional, default} \code{NULL}) \cr
A \code{vector} containing pseudo-absence set to be loaded, must be among \code{PA1},
\code{PA2}, \code{...}, \code{allData}}

\item{run}{(\emph{optional, default} \code{NULL}) \cr
A \code{vector} containing repetition set to be loaded, must be among \code{RUN1},
\code{RUN2}, \code{...}, \code{allRun}}

\item{algo}{(\emph{optional, default} \code{NULL}) \cr
A \code{character} containing algorithm to be loaded, must be either
\code{ANN}, \code{CTA}, \code{FDA}, \code{GAM}, \code{GBM}, \code{GLM}, \code{MARS},
\code{MAXENT}, \code{MAXNET}, \code{RF}, \code{SRE}, \code{XGBOOST}}

\item{merged.by.PA}{(\emph{optional, default} \code{NULL}) \cr
A \code{vector} containing merged pseudo-absence set to be loaded, must be among \code{PA1},
\code{PA2}, \code{...}, \code{mergedData}}

\item{merged.by.run}{(\emph{optional, default} \code{NULL}) \cr
A \code{vector} containing merged repetition set to be loaded, must be among \code{RUN1},
\code{RUN2}, \code{...}, \code{mergedRun}}

\item{merged.by.algo}{(\emph{optional, default} \code{NULL}) \cr
A \code{character} containing merged algorithm to be loaded, must be among
\code{ANN}, \code{CTA}, \code{FDA}, \code{GAM}, \code{GBM}, \code{GLM}, \code{MARS},
\code{MAXENT}, \code{MAXNET}, \code{RF}, \code{SRE}, \code{XGBOOST}, \code{mergedAlgo}}

\item{filtered.by}{(\emph{optional, default} \code{NULL}) \cr
A \code{vector} containing evaluation metric selected to filter single models to build the
ensemble models, must be among \code{POD}, \code{FAR}, \code{POFD}, \code{SR},
\code{ACCURACY}, \code{BIAS}, \code{ROC}, \code{TSS}, \code{KAPPA}, \code{OR}, \code{ORSS},
\code{CSI}, \code{ETS}, \code{BOYCE}, \code{MPA}}

\item{metric.eval}{(\emph{optional, default} \code{NULL}) \cr
A \code{vector} containing evaluation metric to be kept, must be among \code{POD},
\code{FAR}, \code{POFD}, \code{SR}, \code{ACCURACY}, \code{BIAS}, \code{ROC}, \code{TSS},
\code{KAPPA}, \code{OR}, \code{ORSS}, \code{CSI}, \code{ETS}, \code{BOYCE}, \code{MPA}}

\item{expl.var}{(\emph{optional, default} \code{NULL}) \cr
A \code{vector} containing explanatory variables to be kept, that can be obtained with the
\code{\link{get_formal_data}(obj, subinfo = 'expl.var.names')} function}

\item{metric.binary}{(\emph{optional, default} \code{NULL}) \cr
A \code{vector} containing evaluation metric selected to transform predictions into binary
values, must be among \code{POD}, \code{FAR}, \code{POFD}, \code{SR}, \code{ACCURACY},
\code{BIAS}, \code{ROC}, \code{TSS}, \code{KAPPA}, \code{OR}, \code{ORSS}, \code{CSI},
\code{ETS}, \code{BOYCE}, \code{MPA}}

\item{metric.filter}{(\emph{optional, default} \code{NULL}) \cr
A \code{vector} containing evaluation metric to filter predictions, must be among \code{POD},
\code{FAR}, \code{POFD}, \code{SR}, \code{ACCURACY}, \code{BIAS}, \code{ROC}, \code{TSS},
\code{KAPPA}, \code{OR}, \code{ORSS}, \code{CSI}, \code{ETS}, \code{BOYCE}, \code{MPA}}

\item{model.as.col}{(\emph{optional, default} \code{FALSE}) \cr
A \code{boolean} given to \code{\link{get_predictions}}. If \code{TRUE}
prediction are returned as a wide \code{data.frame} with each column containing
predictions for a single model and corresponding to the old output given by
\pkg{biomod2} in version < 4.2-2. If \code{FALSE} predictions are returned
as a long \code{data.frame} with many additional informations readily
available.}

\item{...}{additional parameters}
}
\description{
\code{MS_Modeling()} output object class

Functions to extract informations from \code{\link{MS.models.out}},
\code{\link{MS.projection.out}} or \code{\link{MS.ensemble.models.out}} objects
}
\section{Slots}{

\describe{
\item{\code{ms.project}}{a \code{character} corresponding to the name of
the multispecies project}

\item{\code{modeling.id}}{a \code{character} corresponding to the name (ID) of the
simulation set}

\item{\code{dir.name}}{a \code{character} corresponding to the modeling folder}

\item{\code{sp.name}}{a \code{character vector} corresponding to the species name}

\item{\code{data.type}}{a \code{character} corresponding to the data type}

\item{\code{expl.var.names}}{a \code{vector} containing names of explanatory
variables}

\item{\code{models.computed}}{a \code{vector} containing names of computed models}

\item{\code{models.failed}}{a \code{vector} containing names of failed models}

\item{\code{has.evaluation.data}}{a \code{logical} value defining whether evaluation
data is given}

\item{\code{scale.models}}{a \code{logical} value defining whether models have been
rescaled or not}

\item{\code{link}}{a \code{character} containing the file name of the saved object}
}}

\author{
Hélène Blancheteau
}
