% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_EnsembleModeling.R
\name{MS_EnsembleModeling}
\alias{MS_EnsembleModeling}
\title{Create and evaluate an ensemble set of models and predictions}
\usage{
MS_EnsembleModeling(
  ms.mod,
  models.chosen = "all",
  em.by = "PA+run",
  em.algo,
  metric.select = "all",
  metric.select.thresh = NULL,
  metric.select.table = NULL,
  metric.select.dataset = NULL,
  metric.eval = c("KAPPA", "TSS", "ROC"),
  var.import = 0,
  EMci.alpha = 0.05,
  EMwmean.decay = "proportional",
  nb.cpu = 1,
  seed.val = NULL
)
}
\arguments{
\item{ms.mod}{a \code{\link{MS.models.out}} object returned by the
\code{\link{MS_Modeling}} function}

\item{models.chosen}{a \code{vector} containing model names to be kept, must be either
\code{all} or a sub-selection of model names that can be obtained with the
\code{\link{get_built_models}} function}

\item{em.by}{a \code{character} corresponding to the way kept models will be combined to build
the ensemble models, must be among \code{all}, \code{algo}, \code{PA}, \code{PA+algo},
\code{PA+run}}

\item{em.algo}{a \code{vector} corresponding to the ensemble models that will be computed,
must be among \code{'EMmean'}, \code{'EMmedian'}, \code{'EMcv'}, \code{'EMci'},
\code{'EMca'}, \code{'EMwmean'}}

\item{metric.select}{a \code{vector} containing evaluation metric names to be used together with
\code{metric.select.thresh} to exclude single models based on their evaluation scores
(for ensemble methods like probability weighted mean or committee averaging). Must be among
\code{all} (same evaluation metrics than those of \code{bm.mod}), \code{user.defined}
(and defined through \code{metric.select.table}) or \code{POD}, \code{FAR}, \code{POFD},
\code{SR}, \code{ACCURACY}, \code{BIAS}, \code{ROC}, \code{TSS}, \code{KAPPA}, \code{OR},
\code{ORSS}, \code{CSI}, \code{ETS}, \code{BOYCE}, \code{MPA}}

\item{metric.select.thresh}{(\emph{optional, default} \code{NULL}) \cr
A \code{vector} of \code{numeric} values corresponding to the minimum scores (one for each
\code{metric.select}) below which single models will be excluded from the ensemble model
building}

\item{metric.select.table}{(\emph{optional, default} \code{NULL}) \cr If
\code{metric.select = 'user.defined'}, a \code{data.frame} containing evaluation scores
calculated for each single models and that will be compared to \code{metric.select.thresh}
values to exclude some of them from the ensemble model building, with \code{metric.select}
rownames, and \code{models.chosen} colnames}

\item{metric.select.dataset}{(\emph{optional, default} \code{'validation'}
\emph{if possible}). A character determining which dataset should be used to filter and/or
weigh the ensemble models should be among 'evaluation', 'validation' or 'calibration'.}

\item{metric.eval}{a \code{vector} containing evaluation metric names to be used, must
be among  \code{POD}, \code{FAR}, \code{POFD}, \code{SR}, \code{ACCURACY}, \code{BIAS},
\code{ROC}, \code{TSS}, \code{KAPPA}, \code{OR}, \code{ORSS}, \code{CSI}, \code{ETS},
\code{BOYCE}, \code{MPA}}

\item{var.import}{(\emph{optional, default} \code{NULL}) \cr
An \code{integer} corresponding to the number of permutations to be done for each variable to
estimate variable importance}

\item{EMci.alpha}{(\emph{optional, default} \code{0.05}) \cr
A \code{numeric} value corresponding to the significance level to estimate confidence interval}

\item{EMwmean.decay}{(\emph{optional, default} \code{proportional}) \cr A
value defining the relative importance of the weights (if \code{'EMwmean'}
was given to argument \code{em.algo}). A high value will strongly
discriminate \emph{good} models from the \emph{bad} ones (see Details),
while \code{proportional} will attribute weights proportionally to the
models evaluation scores}

\item{nb.cpu}{(\emph{optional, default} \code{1}) \cr
An \code{integer} value corresponding to the number of computing resources to be used to
parallelize the single models predictions and the ensemble models computation}

\item{seed.val}{(\emph{optional, default} \code{NULL}) \cr
An \code{integer} value corresponding to the new seed value to be set}
}
\value{
A \code{\link{MS.ensemble.models.out}} object acting as a proxi for the created \code{BIOMOD.ensemble.models.out}.
}
\description{
This function allows to combine a range of models built with the
\code{\link{MS_Modeling}} function in one (or several) ensemble model. Modeling
uncertainty can be assessed as well as variables importance, ensemble predictions can be
evaluated against original data, and created ensemble models can be projected over new
conditions (see Details).
}
\examples{
library(terra)
library(biomod2)

# Load species occurrences (6 species available)
data(DataSpecies)

# Select the name of the studied species
myRespName <- c("PantheraOnca", "PteropusGiganteus")

# Get corresponding presence/absence data
myResp <- DataSpecies[, myRespName]

# Get corresponding XY coordinates
myRespXY <- DataSpecies[, c('X_WGS84', 'Y_WGS84')]

# Load environmental variables extracted from BIOCLIM (bio_3, bio_4, bio_7, bio_11 & bio_12)
data(bioclim_current)
myExpl <- terra::rast(bioclim_current)



myMSData <- MS_FormatingData(ms.project.name = "Example_MS",
                             resp.name = myRespName,
                             resp.var = myResp,
                             expl.var = myExpl,
                             data.type = "binary",
                             resp.xy = myRespXY)

params.CV <- list("PantheraOnca" = list(CV.strategy = "random", CV.nb.rep = 2, CV.perc = 0.8),
                  "PteropusGiganteus" = list(CV.strategy = "random", CV.nb.rep = 2, CV.perc = 0.8))

params.OPT <- list("PantheraOnca" = list(OPT.strategy = "bigboss"),
                   "PteropusGiganteus" = list(OPT.strategy = "bigboss"))

myMSModelOut <- MS_Modeling(myMSData, 
                            modeling.id = "FirstModels",
                            models = c("GLM", "XGBOOST"),
                            params.CV = params.CV,
                            params.OPT = params.OPT)

metric.select.tresh.byspecies <- list("PantheraOnca" = 0.5,
                                      "PteropusGiganteus" = 0.4)

myMSEM <- MS_EnsembleModeling(myMSModelOut,
                              models.chosen = 'all',
                              em.by = 'all',
                              em.algo = c("EMmean", "EMca"),
                              metric.select = 'TSS',
                              metric.select.thresh = metric.select.tresh.byspecies,
                              metric.eval = c('TSS', 'ROC'))

myMSEM

get_evaluations(myMSEM, sp = "PantheraOnca")


\dontshow{
  unlink('Example_MS', recursive = TRUE)
}

}
\seealso{
\code{\link{BIOMOD_FormatingData}}, \code{\link{bm_ModelingOptions}},
\code{\link{bm_CrossValidation}}, \code{\link{bm_VariablesImportance}},
\code{\link{BIOMOD_Modeling}}, \code{\link{BIOMOD_EnsembleForecasting}},
\code{\link{bm_PlotEvalMean}}, \code{\link{bm_PlotEvalBoxplot}},
\code{\link{bm_PlotVarImpBoxplot}}, \code{\link{bm_PlotResponseCurves}}

Other Main functions: 
\code{\link{MS_EnsembleForecasting}()},
\code{\link{MS_FormatingData}()},
\code{\link{MS_Projection}()}
}
\author{
Helene Blancheteau
}
\concept{Main functions}
\keyword{ensemble}
\keyword{models}
\keyword{weights}
