% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_FormatingData.R
\name{MS_FormatingData}
\alias{MS_FormatingData}
\title{Format input data for usage in \pkg{biomod3}}
\usage{
MS_FormatingData(
  ms.project.name,
  dir.name = ".",
  resp.name = NULL,
  resp.var = NULL,
  expl.var = NULL,
  data.type = NULL,
  resp.xy = NULL,
  eval.resp.var = NULL,
  eval.expl.var = NULL,
  eval.resp.xy = NULL,
  params.PA = NULL,
  single.formated.data = NULL,
  filter.raster = FALSE,
  seed.val = NULL
)
}
\arguments{
\item{ms.project.name}{a \code{character} corresponding to the name of your project for your multispecies modelling}

\item{dir.name}{(\emph{optional, default} \code{.}) \cr
A \code{character} corresponding to the modeling folder}

\item{resp.name}{a \code{character vector} corresponding to the species name}

\item{resp.var}{a \code{vector} or a \code{\link[terra:vect]{SpatVector}} containing your response variable (See Details).}

\item{expl.var}{a \code{matrix}, \code{data.frame}, \code{\link[terra:vect]{SpatVector}}
or \code{\link[terra:rast]{SpatRaster}} object containing the explanatory variables
(in columns or layers) that will be used to build the species distribution model(s)
\cr \emph{Note that old format from \pkg{raster} and \pkg{sp} are still supported such as
\code{RasterStack} and \code{SpatialPointsDataFrame} objects. }}

\item{data.type}{a \code{character}, corresponding to the response data type to be used, must be either
\code{binary}, \code{count}, \code{ordinal}, \code{relative}, or \code{abundance}. If data.type is not provided,
\code{biomod2} will try to guess.}

\item{resp.xy}{(\emph{optional, default} \code{NULL}) \cr
If \code{resp.var} is a \code{vector}, a 2-columns \code{matrix} or \code{data.frame}
containing the corresponding \code{X} and \code{Y} coordinates that will be
used to build the species distribution model(s)}

\item{eval.resp.var}{(\emph{optional, default} \code{NULL}) \cr
A \code{vector}, a \code{\link[terra:vect]{SpatVector}}
without associated data (\emph{if presence-only}),
or a \code{\link[terra:vect]{SpatVector}} object containing binary data
(\code{0} : absence, \code{1} : presence, \code{NA} : indeterminate)
for a single species that will be used to
evaluate the species distribution model(s) with independent data
\cr \emph{Note that old format from \pkg{sp} are still supported such as
\code{SpatialPoints}  (if presence-only) or \code{SpatialPointsDataFrame}
object containing binary data.}}

\item{eval.expl.var}{(\emph{optional, default} \code{NULL}) \cr
A \code{matrix}, \code{data.frame}, \code{\link[terra:vect]{SpatVector}}
or \code{\link[terra:rast]{SpatRaster}} object containing the explanatory variables (in
columns or layers) that will be used to evaluate the species distribution model(s) with
independent data.
\cr \emph{Note that old format from \pkg{raster} and \pkg{sp} are still supported such as
\code{RasterStack} and \code{SpatialPointsDataFrame} objects. }}

\item{eval.resp.xy}{(\emph{optional, default} \code{NULL}) \cr
If \code{resp.var} is a \code{vector}, a 2-columns \code{matrix} or \code{data.frame}
containing the corresponding \code{X} and \code{Y} coordinates that will be used to evaluate
the species distribution model(s) with independent data}

\item{params.PA}{a \code{list} with the species names associated to the parameters of PA}

\item{single.formated.data}{a \code{list} of \code{BIOMOD.formated.data} objects to add to the MS.formated.data object}

\item{filter.raster}{(\emph{optional, default} \code{FALSE}) \cr
If \code{expl.var} is of raster type, a \code{logical} value defining whether \code{resp.var}
is to be filtered when several points occur in the same raster cell}

\item{seed.val}{(\emph{optional, default} \code{NULL}) \cr
An \code{integer} value corresponding to the new seed value to be set}
}
\value{
A \code{\link{MS.formated.data}} object that can be used to build species distribution
model(s) with the \code{\link{MS_Modeling}} function. \cr
}
\description{
This function gathers together all input data needed (\emph{xy,
presences/absences, explanatory variables, and the same for evaluation data if available}) to
run \pkg{biomod2} models. It run the function BIOMOD_FormatingData for all the different species.
}
\examples{
library(terra)
library(biomod2)

# Load species occurrences (6 species available)
data(DataSpecies)
head(DataSpecies)

# Select the name of the studied species
myRespName <- c("PantheraOnca", "PteropusGiganteus")

# Get corresponding presence/absence data
myResp <- DataSpecies[, myRespName]

# Get corresponding XY coordinates
myRespXY <- DataSpecies[, c('X_WGS84', 'Y_WGS84')]

# Load environmental variables extracted from BIOCLIM (bio_3, bio_4, bio_7, bio_11 & bio_12)
data(bioclim_current)
myExpl <- terra::rast(bioclim_current)

\dontshow{
  myExtent <- terra::ext(0,30,45,70)
  myExpl <- terra::crop(myExpl, myExtent)
}



myMSData <- MS_FormatingData(ms.project.name = "Example_MS",
                             resp.name = myRespName,
                             resp.var = myResp,
                             expl.var = myExpl,
                             data.type = "binary",
                             resp.xy = myRespXY)

myMSData
summary(myMSData)
plot(myMSData)

\dontshow{
  unlink('Example_MS', recursive = TRUE)
}




}
\seealso{
\code{\link{bm_PseudoAbsences}}, \code{\link{BIOMOD_Modeling}}

Other Main functions: 
\code{\link{MS_EnsembleForecasting}()},
\code{\link{MS_EnsembleModeling}()},
\code{\link{MS_Projection}()}
}
\author{
Helene Blancheteau
}
\concept{Main functions}
\keyword{dataset}
\keyword{evaluation}
\keyword{format}
\keyword{pseudo-absence}
