% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_Modeling.R
\name{MS_Modeling}
\alias{MS_Modeling}
\title{Run a range of species distribution models}
\usage{
MS_Modeling(
  ms.format,
  modeling.id = as.character(format(Sys.time(), "\%s")),
  models = c("ANN", "CTA", "FDA", "GAM", "GBM", "GLM", "MARS", "MAXNET", "RF", "RFd",
    "SRE", "XGBOOST"),
  params.CV = NULL,
  params.OPT = NULL,
  weights = NULL,
  prevalence = NULL,
  metric.eval = c("KAPPA", "TSS", "ROC"),
  var.import = 0,
  scale.models = FALSE,
  nb.cpu = 1,
  seed.val = NULL
)
}
\arguments{
\item{ms.format}{a \code{\link{MS.formated.data}} object returned by the \code{\link{BIOMOD_FormatingData}} function}

\item{modeling.id}{a \code{character} corresponding to the name (ID) of the simulation set
(\emph{a random number by default})}

\item{models}{a \code{vector} containing model names to be computed, must be among
\code{ANN}, \code{CTA}, \code{FDA}, \code{GAM}, \code{GBM}, \code{GLM}, \code{MARS},
\code{MAXENT}, \code{MAXNET}, \code{RF}, \code{RFd}, \code{SRE}, \code{XGBOOST}}

\item{params.CV}{a \code{list} with the species names associated to the parameters of Cross-Validation. See BIOMOD_Modeling}

\item{params.OPT}{a \code{list} with the species names associated to the options of the algorithms. See BIOMOD_Modeling}

\item{weights}{(\emph{optional, default} \code{NULL}) \cr
A \code{vector} of \code{numeric} values corresponding to observation weights (one per
observation, see Details)}

\item{prevalence}{(\emph{optional, default} \code{NULL}) \cr
A \code{numeric} between \code{0} and \code{1} corresponding to the species prevalence to
build '\emph{weighted response weights}' (see Details)}

\item{metric.eval}{a \code{vector} containing evaluation metric names to be used, must
be among \code{ROC}, \code{TSS}, \code{KAPPA}, \code{ACCURACY}, \code{BIAS}, \code{POD},
\code{FAR}, \code{POFD}, \code{SR}, \code{CSI}, \code{ETS}, \code{OR},
\code{ORSS}, \code{BOYCE}, \code{MPA}, \code{RMSE}, \code{MAE}, \code{MSE}, \code{Rsquared}, \code{Rsquared_aj},
\code{Max_error}, \code{Accuracy}, \code{"Recall"}, \code{"Precision"}, \code{"F1"}}

\item{var.import}{(\emph{optional, default} \code{NULL}) \cr
An \code{integer} corresponding to the number of permutations to be done for each variable to
estimate variable importance}

\item{scale.models}{(\emph{optional, default} \code{FALSE}) \cr
A \code{logical} value defining whether all models predictions should be scaled with a
binomial GLM or not}

\item{nb.cpu}{(\emph{optional, default} \code{1}) \cr
An \code{integer} value corresponding to the number of computing resources to be used to
parallelize the single models computation}

\item{seed.val}{(\emph{optional, default} \code{NULL}) \cr
An \code{integer} value corresponding to the new seed value to be set}
}
\value{
A \code{\link{MS.models.out}} object acting as a proxi for the created \code{BIOMOD.models.out}.
}
\description{
This function allows to calibrate and evaluate a range of modeling techniques
for a given species distribution. The dataset can be split up in calibration/validation parts,
and the predictive power of the different models can be estimated using a range of evaluation
metrics (see Details).
}
\author{
Helene Blancheteau
}
