% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomod3_class1_Fd.R
\docType{methods}
\name{plot,MS.formated.data,missing-method}
\alias{plot,MS.formated.data,missing-method}
\title{\code{plot} method for \code{\link{MS.formated.data}} object class}
\usage{
\S4method{plot}{MS.formated.data,missing}(x, plot.type, plot.output, plot.eval, point.size = 1.5, do.plot = TRUE)
}
\arguments{
\item{x}{a \code{\link{MS.formated.data}} object. Coordinates must be available to be able to use \code{plot}.}

\item{plot.type}{a \code{character}, either \code{'points'} (\emph{default})
or \code{'raster'} (\emph{if environmental variables were given as a raster}).
With \code{plot.type = 'points'} occurrences will be represented as points
(better when using fine-grained data). With \code{plot.type = 'raster'}
occurrences will be represented as a raster (better when using coarse-grained
data)}

\item{plot.output}{a \code{character}, either \code{'facet'} (\emph{default})
or \code{'list'}. \code{plot.output} determines whether plots are returned
as a single facet with all plots or a \code{list} of individual plots
(better when there are numerous graphics)}

\item{plot.eval}{(\emph{optional, default} \code{TRUE}) \cr
A \code{logical} defining whether evaluation data should be added to the plot or not}

\item{point.size}{a \code{numeric} to adjust the size of points when
\code{plot.type = 'points'}.}

\item{do.plot}{(\emph{optional, default} \code{TRUE}) \cr
A \code{logical} defining whether the plot is to be rendered or not}
}
\value{
a \code{list} with the data used to generate the plot and a
\code{ggplot2} object
}
\description{
Plot the spatial distribution of presences, absences and
pseudo-absences among the different potential dataset (calibration,
validation and evaluation). Available only if coordinates were given to
\code{\link{MS_FormatingData}}.
}
