% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomod3_class5_Wrap.R
\docType{class}
\name{BIOMOD.wrap.out}
\alias{BIOMOD.wrap.out}
\alias{BIOMOD.wrap.out-class}
\alias{show,BIOMOD.wrap.out-method}
\title{\code{BIOMOD_Wrap()} output object class}
\usage{
\S4method{show}{BIOMOD.wrap.out}(object)
}
\arguments{
\item{object}{a \code{\link{BIOMOD.wrap.out}} object}
}
\description{
Class returned by \code{\link{BIOMOD_Wrap}}, and used by
\code{\link{BIOMOD_ProjectionWrap}}
}
\section{Slots}{

\describe{
\item{\code{formated.data}}{a \code{BIOMOD.formated.data} object}

\item{\code{single.models}}{a \code{BIOMOD.models.out} object}

\item{\code{ensemble.models}}{a \code{BIOMOD.ensemble.models.out} object}

\item{\code{output}}{a \code{character} link to the output file}
}}

\author{
Helene Blancheteau
}
