% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomod3_class2_SM.R, R/biomod3_class3_EM.R,
%   R/biomod3_class4_Proj.R
\name{getters.out}
\alias{getters.out}
\alias{get_species_data}
\alias{get_eval_data}
\alias{get_options}
\alias{get_calib_lines}
\alias{get_formal_data}
\alias{get_projected_models}
\alias{free}
\alias{get_predictions}
\alias{get_kept_models}
\alias{get_built_models}
\alias{get_evaluations}
\alias{get_variables_importance}
\alias{get_options,MS.models.out-method}
\alias{get_calib_lines,MS.models.out-method}
\alias{get_formal_data,MS.models.out-method}
\alias{get_predictions,MS.models.out-method}
\alias{get_built_models,MS.models.out-method}
\alias{get_evaluations,MS.models.out-method}
\alias{get_variables_importance,MS.models.out-method}
\alias{get_formal_data,MS.ensemble.models.out-method}
\alias{get_built_models,MS.ensemble.models.out-method}
\alias{get_kept_models,MS.ensemble.models.out-method}
\alias{get_predictions,MS.ensemble.models.out-method}
\alias{get_evaluations,MS.ensemble.models.out-method}
\alias{get_variables_importance,MS.ensemble.models.out-method}
\alias{get_projected_models,MS.projection.out-method}
\alias{get_predictions,MS.projection.out-method}
\title{Functions to extract informations from \code{\link{MS.models.out}},
\code{\link{MS.projection.out}} or \code{\link{MS.ensemble.models.out}} objects}
\usage{
get_species_data(obj, ...)

\S4method{get_options}{MS.models.out}(obj, sp)

\S4method{get_calib_lines}{MS.models.out}(obj, as.data.frame = FALSE, PA = NULL, run = NULL)

\S4method{get_formal_data}{MS.models.out}(obj, sp, subinfo = NULL)

\S4method{get_predictions}{MS.models.out}(
  obj,
  sp,
  evaluation = FALSE,
  full.name = NULL,
  PA = NULL,
  run = NULL,
  algo = NULL,
  model.as.col = FALSE
)

\S4method{get_built_models}{MS.models.out}(obj, full.name = NULL, PA = NULL, run = NULL, algo = NULL)

\S4method{get_evaluations}{MS.models.out}(
  obj,
  sp,
  full.name = NULL,
  PA = NULL,
  run = NULL,
  algo = NULL,
  metric.eval = NULL
)

\S4method{get_variables_importance}{MS.models.out}(
  obj,
  sp,
  full.name = NULL,
  PA = NULL,
  run = NULL,
  algo = NULL,
  expl.var = NULL
)

\S4method{get_formal_data}{MS.ensemble.models.out}(obj, sp, subinfo = NULL)

\S4method{get_built_models}{MS.ensemble.models.out}(
  obj,
  full.name = NULL,
  merged.by.algo = NULL,
  merged.by.run = NULL,
  merged.by.PA = NULL,
  filtered.by = NULL,
  algo = NULL
)

\S4method{get_kept_models}{MS.ensemble.models.out}(obj)

\S4method{get_predictions}{MS.ensemble.models.out}(
  obj,
  sp,
  evaluation = FALSE,
  full.name = NULL,
  merged.by.algo = NULL,
  merged.by.run = NULL,
  merged.by.PA = NULL,
  filtered.by = NULL,
  algo = NULL,
  model.as.col = FALSE
)

\S4method{get_evaluations}{MS.ensemble.models.out}(
  obj,
  sp,
  full.name = NULL,
  merged.by.algo = NULL,
  merged.by.run = NULL,
  merged.by.PA = NULL,
  filtered.by = NULL,
  algo = NULL,
  metric.eval = NULL
)

\S4method{get_variables_importance}{MS.ensemble.models.out}(
  obj,
  sp,
  full.name = NULL,
  merged.by.algo = NULL,
  merged.by.run = NULL,
  merged.by.PA = NULL,
  filtered.by = NULL,
  algo = NULL,
  expl.var = NULL
)

\S4method{get_projected_models}{MS.projection.out}(
  obj,
  full.name = NULL,
  PA = NULL,
  run = NULL,
  algo = NULL,
  merged.by.algo = NULL,
  merged.by.run = NULL,
  merged.by.PA = NULL,
  filtered.by = NULL
)

\S4method{get_predictions}{MS.projection.out}(
  obj,
  sp,
  metric.binary = NULL,
  metric.filter = NULL,
  full.name = NULL,
  PA = NULL,
  run = NULL,
  algo = NULL,
  merged.by.algo = NULL,
  merged.by.run = NULL,
  merged.by.PA = NULL,
  filtered.by = NULL,
  model.as.col = FALSE,
  ...
)
}
\description{
Functions to extract informations from \code{\link{MS.models.out}},
\code{\link{MS.projection.out}} or \code{\link{MS.ensemble.models.out}} objects
}
